#163 Adjust precision so ms values fit

+++ esp8266/src/clacks.h
@@ -9,35 +9,42 @@
 #define CLACKS_MQTT_TOPIC "powerpi/config/%s/change"
 
 // general defaults
+// the delay between sensor polling (half a second)
 #define POLL_DELAY 0.5
 
 // defaults for DHT22
+// the number of seconds between polls (5 minutes)
 #define DHT22_SKIP 5 * 60
 
 // defaults for PIR
+// the number of seconds to allow the PIR to initialise (60s)
 #define PIR_INIT_DELAY 60
+
+// the number of seconds to skip between a transition (5s)
 #define PIR_POST_DETECT_SKIP 5
+
+// the number of seconds to skip after detection (20s)
 #define PIR_POST_MOTION_SKIP 20
 
 struct ClacksConfig_s {
     // whether the configuration was received yet
     bool received;
 
-    // the delay between sensor polling (half a second)
-    unsigned short pollDelay;
+    // the delay between sensor polling in milliseconds
+    unsigned int pollDelay;
 
     // options for the DHT22 sensor
-    // the number of seconds between polls (5 minutes)
+    // the number of intervals to skip between readings
     unsigned short dht22Skip;
 
     // options for the PIR sensor
-    // the number of seconds to allow the PIR to initialise (60s)
-    unsigned short pirInitDelay;
+    // the number of milliseconds to allow the PIR to initialise
+    unsigned int pirInitDelay;
 
-    // the number of seconds to skip between a transition (5s)
+    // the number of intervals to skip between a transition
     unsigned short pirPostDetectSkip;
 
-    // the number of seconds to skip after detection (20s)
+    // the number of intervals to skip after detection
     unsigned short pirPostMotionSkip;
 } ClacksConfig_default = {
     false,
